refactor(graphql-gateway): modularize server, schema, and resolvers
Move server configuration, type definitions, and resolvers into separate modules to improve code organization and maintainability. This change simplifies the main index.js file and enhances readability.
